refactor: rename AuthenticationRepository to UserRepository and create UserRepository interface
