+++ lib/data/repositories/authentication_repository_impl.dart
@@ -5,7 +5,7 @@ import 'package:on_time_front/data/data_sources/authentication_remote_data_sourc
 import 'package:on_time_front/data/data_sources/token_local_data_source.dart';
 import 'package:on_time_front/data/models/sign_in_with_google_request_model.dart';
 import 'package:on_time_front/domain/entities/user_entity.dart';
-import 'package:on_time_front/domain/repositories/authentication_repository.dart';
+import 'package:on_time_front/domain/repositories/user_repository.dart';
 import 'package:rxdart/subjects.dart';
 
 @Singleton(as: UserRepository)
